Improvements to refrigerator opening/closing.
The refrigerator is not closed automatically when Rowena moves. And we
can continue to pick objects outside the refrigerator even when the
refrigerator is open. By picking the refrigerator door, the user can
explicitly close the refrigerator.

If Rowena does something else while the refrigerator is open (e.g. jumps
out the window), she first walks back to the refrigerator and closes it,
if she is close enough. If she is too far away (auto_close_distance),
the refrigerator just closes itself.