Ignore stuff behind open fridge door (issue #10).
Background colliders now have an optional "bg_level" metadata value.
Objects with a higher value are considered to be on top of objects with
a lower value. By default, objects are at level 0. The open refrigerator
door is at level 1, so objects behind it are ignored. And things inside
the refrigerator are at level 2.

The collider for the open refrigerator door is now a simple rectangle.