Exported more timeout properties.
If the user uses an object from the inventory and then move the mouse
outside the box, "Inventory Timeout" is how long before it is closed
automatically.

Comment timeouts now vary with the length of the comment: "Comment
Timeout Base", plus "Comment Timeout per Letter" multiplied by the
number of letters in the comment.